requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
autopep8==1.4.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0
django-allauth==0.42.0
django-crispy-forms==1.9.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
pep8==1.7.1
progress==1.5
pycodestyle==2.5.0
pylint==2.5.3
pyparsing==2.4.6
python3-openid==3.1.0
pytoml==0.1.21
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
sqlparse==0.2.4
toml==0.10.1
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
